requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
h11
